E2e for both new and old apis
- controller linked to new openai service only
