+++ sample-code/spring-app/src/main/java/com/sap/ai/sdk/app/services/OpenAiService.java
@@ -3,33 +3,18 @@ package com.sap.ai.sdk.app.services;
 import static com.sap.ai.sdk.foundationmodels.openai.OpenAiModel.GPT_35_TURBO;
 import static com.sap.ai.sdk.foundationmodels.openai.OpenAiModel.GPT_4O;
 import static com.sap.ai.sdk.foundationmodels.openai.OpenAiModel.TEXT_EMBEDDING_ADA_002;
-import static com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionRequestMessageContentPartImage.TypeEnum.IMAGE_URL;
-import static com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionRequestMessageContentPartImageImageUrl.DetailEnum.HIGH;
-import static com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionRequestMessageContentPartText.TypeEnum.TEXT;
-import static com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionRequestUserMessage.RoleEnum.USER;
-import static com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionTool.TypeEnum.*;
+import static com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionTool.ToolType.FUNCTION;
 
 import com.sap.ai.sdk.core.AiCoreService;
-import com.sap.ai.sdk.foundationmodels.openai.OpenAiChatCompletionDelta;
-import com.sap.ai.sdk.foundationmodels.openai.OpenAiChatCompletionRequest;
-import com.sap.ai.sdk.foundationmodels.openai.OpenAiChatCompletionResponse;
 import com.sap.ai.sdk.foundationmodels.openai.OpenAiClient;
-import com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionNamedToolChoice;
-import com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionNamedToolChoiceFunction;
-import com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionRequestMessageContentPartImage;
-import com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionRequestMessageContentPartImageImageUrl;
-import com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionRequestMessageContentPartText;
-import com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionRequestUserMessage;
-import com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionRequestUserMessageContent;
-import com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionTool;
-import com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionToolChoiceOption;
-import com.sap.ai.sdk.foundationmodels.openai.model2.CreateChatCompletionRequest;
-import com.sap.ai.sdk.foundationmodels.openai.model2.CreateChatCompletionResponse;
-import com.sap.ai.sdk.foundationmodels.openai.model2.EmbeddingsCreate200Response;
-import com.sap.ai.sdk.foundationmodels.openai.model2.EmbeddingsCreateRequest;
-import com.sap.ai.sdk.foundationmodels.openai.model2.EmbeddingsCreateRequestInput;
-import com.sap.ai.sdk.foundationmodels.openai.model2.FunctionObject;
-import java.net.URI;
+import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionDelta;
+import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionFunction;
+import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionOutput;
+import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionParameters;
+import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionTool;
+import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatMessage;
+import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiEmbeddingOutput;
+import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiEmbeddingParameters;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
@@ -49,7 +34,7 @@ public class OpenAiService {
    * @return the assistant message response
    */
   @Nonnull
-  public OpenAiChatCompletionResponse chatCompletion(@Nonnull final String prompt) {
+  public OpenAiChatCompletionOutput chatCompletion(@Nonnull final String prompt) {
     return OpenAiClient.forModel(GPT_35_TURBO).chatCompletion(prompt);
   }
 
@@ -62,19 +47,8 @@ public class OpenAiService {
   public Stream<OpenAiChatCompletionDelta> streamChatCompletionDeltas(
       @Nonnull final String message) {
     final var request =
-        new CreateChatCompletionRequest()
-            .addMessagesItem(
-                new ChatCompletionRequestUserMessage()
-                    .role(USER)
-                    .content(
-                        ChatCompletionRequestUserMessageContent.create(
-                            List.of(
-                                new ChatCompletionRequestMessageContentPartText()
-                                    .text(message)
-                                    .type(TEXT)))))
-            .tools(null)
-            .functions(null)
-            .parallelToolCalls(null);
+        new OpenAiChatCompletionParameters()
+            .addMessages(new OpenAiChatMessage.OpenAiChatUserMessage().addText(message));
 
     return OpenAiClient.forModel(GPT_35_TURBO).streamChatCompletionDeltas(request);
   }
@@ -98,27 +72,15 @@ public class OpenAiService {
    * @return the assistant message response
    */
   @Nonnull
-  public CreateChatCompletionResponse chatCompletionImage(@Nonnull final String linkToImage) {
-    final var partText =
-        new ChatCompletionRequestMessageContentPartText()
-            .type(TEXT)
-            .text("Describe the following image.");
-    final var partImageUrl =
-        new ChatCompletionRequestMessageContentPartImageImageUrl()
-            .url(URI.create(linkToImage))
-            .detail(HIGH);
-    final var partImage =
-        new ChatCompletionRequestMessageContentPartImage().type(IMAGE_URL).imageUrl(partImageUrl);
-    final var userMessage =
-        new ChatCompletionRequestUserMessage()
-            .role(USER)
-            .content(ChatCompletionRequestUserMessageContent.create(List.of(partText, partImage)));
+  public OpenAiChatCompletionOutput chatCompletionImage(@Nonnull final String linkToImage) {
     final var request =
-        new CreateChatCompletionRequest()
-            .addMessagesItem(userMessage)
-            .functions(null)
-            .tools(null)
-            .parallelToolCalls(null);
+        new OpenAiChatCompletionParameters()
+            .addMessages(
+                new OpenAiChatMessage.OpenAiChatUserMessage()
+                    .addText("Describe the following image.")
+                    .addImage(
+                        linkToImage,
+                        OpenAiChatMessage.OpenAiChatUserMessage.ImageDetailLevel.HIGH));
 
     return OpenAiClient.forModel(GPT_4O).chatCompletion(request);
   }
@@ -126,31 +88,25 @@ public class OpenAiService {
   /**
    * Chat request to OpenAI with a tool.
    *
-   * @param description of the function to be sent to the assistant
+   * @param prompt The prompt to send to the assistant
    * @return the assistant message response
    */
   @Nonnull
-  public OpenAiChatCompletionResponse chatCompletionTools(@Nonnull final String description) {
-    var function =
-        new FunctionObject()
-            .name("fibonacci")
-            .description(description)
-            .parameters(
-                Map.of("type", "object", "properties", Map.of("N", Map.of("type", "integer"))));
-
-    var tool = new ChatCompletionTool().type(FUNCTION).function(function);
-
-    var toolChoice =
-        ChatCompletionToolChoiceOption.create(
-            new ChatCompletionNamedToolChoice()
-                .type(ChatCompletionNamedToolChoice.TypeEnum.FUNCTION)
-                .function(new ChatCompletionNamedToolChoiceFunction().name("fibonacci")));
-
-    var request =
-        new OpenAiChatCompletionRequest(
-                "A pair of rabbits is placed in a field. Each month, every pair produces one new pair, starting from the second month. How many rabbits will there be after 12 months?")
-            .tools(List.of(tool))
-            .toolChoice(toolChoice);
+  public OpenAiChatCompletionOutput chatCompletionTools(@Nonnull final String prompt) {
+    final var question =
+        "A pair of rabbits is placed in a field. Each month, every pair produces one new pair, starting from the second month. How many rabbits will there be after 12 months?";
+    final var par = Map.of("type", "object", "properties", Map.of("N", Map.of("type", "integer")));
+    final var function =
+        new OpenAiChatCompletionFunction()
+            .setName("fibonacci")
+            .setDescription(prompt)
+            .setParameters(par);
+    final var tool = new OpenAiChatCompletionTool().setType(FUNCTION).setFunction(function);
+    final var request =
+        new OpenAiChatCompletionParameters()
+            .addMessages(new OpenAiChatMessage.OpenAiChatUserMessage().addText(question))
+            .setTools(List.of(tool))
+            .setToolChoiceFunction("fibonacci");
 
     return OpenAiClient.forModel(GPT_35_TURBO).chatCompletion(request);
   }
@@ -162,9 +118,8 @@ public class OpenAiService {
    * @return the embedding response
    */
   @Nonnull
-  public EmbeddingsCreate200Response embedding(@Nonnull final String input) {
-    final var request =
-        new EmbeddingsCreateRequest().input(EmbeddingsCreateRequestInput.create(input));
+  public OpenAiEmbeddingOutput embedding(@Nonnull final String input) {
+    final var request = new OpenAiEmbeddingParameters().setInput(input);
 
     return OpenAiClient.forModel(TEXT_EMBEDDING_ADA_002).embedding(request);
   }
@@ -177,7 +132,7 @@ public class OpenAiService {
    * @return the assistant message response
    */
   @Nonnull
-  public OpenAiChatCompletionResponse chatCompletionWithResource(
+  public OpenAiChatCompletionOutput chatCompletionWithResource(
       @Nonnull final String resourceGroup, @Nonnull final String prompt) {
 
     final var destination =
