+++ sample-code/spring-app/src/test/java/com/sap/ai/sdk/app/controllers/NewOpenAiTest.java
@@ -1,34 +1,37 @@
 package com.sap.ai.sdk.app.controllers;
 
 import static com.sap.ai.sdk.foundationmodels.openai.OpenAiModel.GPT_35_TURBO;
+import static com.sap.ai.sdk.foundationmodels.openai.model2.ChatCompletionResponseMessageRole.ASSISTANT;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.sap.ai.sdk.app.services.OpenAiService;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.sap.ai.sdk.app.services.NewOpenAiService;
+import com.sap.ai.sdk.foundationmodels.openai.OpenAiChatCompletionRequest;
 import com.sap.ai.sdk.foundationmodels.openai.OpenAiClient;
-import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionOutput;
-import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionParameters;
-import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatMessage.OpenAiChatUserMessage;
+import com.sap.ai.sdk.foundationmodels.openai.OpenAiMessage;
+import com.sap.ai.sdk.foundationmodels.openai.model2.CompletionUsage;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicReference;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 @Slf4j
-class OpenAiTest {
-  OpenAiService service;
+class NewOpenAiTest {
+  NewOpenAiService service;
 
   @BeforeEach
   void setUp() {
-    service = new OpenAiService();
+    service = new NewOpenAiService();
   }
 
   @Test
   void chatCompletion() {
-    final var completion = service.chatCompletion("Who is the prettiest");
+    final var completion =
+        service.chatCompletion("Who is the prettiest");
 
-    final var message = completion.getChoices().get(0).getMessage();
-    assertThat(message.getRole()).isEqualTo("assistant");
-    assertThat(message.getContent()).isNotEmpty();
+    assertThat(completion.getChoice().getMessage().getRole()).isEqualTo(ASSISTANT);
+    assertThat(completion.getContent()).isNotEmpty();
   }
 
   @Test
@@ -38,24 +41,24 @@ class OpenAiTest {
             "https://upload.wikimedia.org/wikipedia/commons/thumb/5/59/SAP_2011_logo.svg/440px-SAP_2011_logo.svg.png");
 
     final var message = completion.getChoices().get(0).getMessage();
-    assertThat(message.getRole()).isEqualTo("assistant");
+    assertThat(message.getRole()).isEqualTo(ASSISTANT);
     assertThat(message.getContent()).isNotEmpty();
   }
 
   @Test
   void streamChatCompletion() {
-    final var request =
-        new OpenAiChatCompletionParameters()
-            .addMessages(new OpenAiChatUserMessage().addText("Who is the prettiest?"));
+    final var userMessage = OpenAiMessage.user("Who is the prettiest?");
+    final var prompt = new OpenAiChatCompletionRequest(userMessage);
 
-    final var totalOutput = new OpenAiChatCompletionOutput();
+    final var totalOutput = new AtomicReference<CompletionUsage>();
     final var filledDeltaCount = new AtomicInteger(0);
     OpenAiClient.forModel(GPT_35_TURBO)
-        .streamChatCompletionDeltas(request)
-        .peek(totalOutput::addDelta)
+        .streamChatCompletionDeltas(prompt)
         // foreach consumes all elements, closing the stream at the end
         .forEach(
             delta -> {
+              final var usage = delta.getCompletionUsage(new ObjectMapper());
+              totalOutput.compareAndExchange(null, usage);
               final String deltaContent = delta.getDeltaContent();
               log.info("delta: {}", delta);
               if (!deltaContent.isEmpty()) {
@@ -67,10 +70,9 @@ class OpenAiTest {
     // see OpenAiChatCompletionDelta#getDeltaContent
     assertThat(filledDeltaCount.get()).isGreaterThan(0);
 
-    assertThat(totalOutput.getChoices()).isNotEmpty();
-    assertThat(totalOutput.getChoices().get(0).getMessage().getContent()).isNotEmpty();
-    assertThat(totalOutput.getPromptFilterResults()).isNotNull();
-    assertThat(totalOutput.getChoices().get(0).getContentFilterResults()).isNotNull();
+    assertThat(totalOutput.get().getTotalTokens()).isGreaterThan(0);
+    assertThat(totalOutput.get().getPromptTokens()).isEqualTo(14);
+    assertThat(totalOutput.get().getCompletionTokens()).isGreaterThan(0);
   }
 
   @Test
@@ -78,8 +80,8 @@ class OpenAiTest {
     final var completion =
         service.chatCompletionTools("Calculate the Fibonacci number for given sequence index.");
 
-    final var message = completion.getChoices().get(0).getMessage();
-    assertThat(message.getRole()).isEqualTo("assistant");
+    final var message = completion.getChoice().getMessage();
+    assertThat(message.getRole()).isEqualTo(ASSISTANT);
     assertThat(message.getToolCalls()).isNotNull();
     assertThat(message.getToolCalls().get(0).getFunction().getName()).isEqualTo("fibonacci");
   }
@@ -98,8 +100,7 @@ class OpenAiTest {
     final var completion =
         service.chatCompletionWithResource("ai-sdk-java-e2e", "Where is the nearest coffee shop?");
 
-    final var message = completion.getChoices().get(0).getMessage();
-    assertThat(message.getRole()).isEqualTo("assistant");
-    assertThat(message.getContent()).isNotEmpty();
+    assertThat(completion.getChoice().getMessage().getRole()).isEqualTo(ASSISTANT);
+    assertThat(completion.getContent()).isNotEmpty();
   }
 }
